ЗП теперь не считается для недовезённых адресов Перенёс запрос энамов из репозитория в сущность
