+++ VodovozBusiness/EntityRepositories/Logistic/IRouteListItemRepository.cs
@@ -13,7 +13,6 @@ namespace Vodovoz.EntityRepositories.Logistic
 		RouteListItem GetTransferedFrom(IUnitOfWork uow, RouteListItem item);
 		bool HasRouteListItemsForOrder(IUnitOfWork uow, Order order);
 		bool WasOrderInAnyRouteList(IUnitOfWork uow, Order order);
-		RouteListItemStatus[] GetUndeliveryStatuses();
 		bool AnotherRouteListItemForOrderExist(IUnitOfWork uow, RouteListItem routeListItem);
 	}
 }
